stream events

+++ status/threads.go
@@ -13,11 +13,11 @@ type ThreadsDetail struct {
 	counter []uint64
 }
 
-func (td *ThreadsDetail) Working(tid uint) {
+func (td *ThreadsDetail) Working(tid uint64) {
 	atomic.StoreUint32(&td.status[tid], 1)
 }
 
-func (td *ThreadsDetail) Idle(tid uint) {
+func (td *ThreadsDetail) Idle(tid uint64) {
 	atomic.StoreUint32(&td.status[tid], 0)
 }
 
@@ -25,11 +25,11 @@ func (td *ThreadsDetail) Status(tid uint) ThreadStatus {
 	return ThreadStatus(atomic.LoadUint32(&td.status[tid]))
 }
 
-func (td *ThreadsDetail) Add(tid uint, n uint64) {
+func (td *ThreadsDetail) Add(tid uint64, n uint64) {
 	atomic.AddUint64(&td.counter[tid], n)
 }
 
-func (td *ThreadsDetail) Count(tid uint) uint64 {
+func (td *ThreadsDetail) Count(tid uint64) uint64 {
 	return atomic.LoadUint64(&td.counter[tid])
 }
 
@@ -41,7 +41,7 @@ func (td *ThreadsDetail) AllCounter() []uint64 {
 	return td.counter
 }
 
-func NewThreadsDetail(total uint) *ThreadsDetail {
+func NewThreadsDetail(total uint64) *ThreadsDetail {
 	td := &ThreadsDetail{
 		status:  make([]uint32, total),
 		counter: make([]uint64, total),
